First attempt, just adding eventhandler
First naive attempt where I just added an eventhandler for keypresses
to the knockout binding. Ended up having to change the status field
to update its value on keydown as well. Perhaps better to trigger
a reevaluation in some other way in case this way is bad from a
performance perspective.